added additional sounds
